Separate payload comparison into its own file
